added Disk abstraction and Mounter class
