+++ src/main/java/org/jenkinsci/plugins/ewm/clouds/AwsEfsMounter.java
@@ -1,0 +1,33 @@
+package org.jenkinsci.plugins.ewm.clouds;
+
+import org.jenkinsci.plugins.ewm.definitions.Disk;
+
+// TODO : do I use a singleton or just static class?
+// TODO : how do I write a static class
+public class AwsEfsMounter {
+    // TODO : how to access the Node and run bashes
+    public static void mountWithIamConfigured() {
+
+    }
+
+    public static void mountWithCredentials() {
+
+    }
+
+
+
+    // TODO : how do we use bash here? Do we use a bash plugin?
+    // TODO : do we need to add parameters at 'mount'
+    private void runBashCommands() {
+
+    }
+
+
+    private void doMount() {
+
+    }
+
+    private void doUnmount() {
+
+    }
+}
